chore: create fromProcess
